move the validation of requests to the version folder
