+++ pkg/apiserver/internal/util/util.go
@@ -17,15 +17,9 @@ limitations under the License.
 package util
 
 import (
-	"fmt"
-
 	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/timestamp"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	v1 "github.com/kubermatic/kubecarrier/pkg/apiserver/api/v1"
 )
 
 func TimestampProto(t *metav1.Time) (*timestamp.Timestamp, error) {
@@ -33,39 +27,4 @@ func TimestampProto(t *metav1.Time) (*timestamp.Timestamp, error) {
 		return nil, nil
 	}
 	return ptypes.TimestampProto(t.Time)
-}
-
-func ValidateGetRequest(req *v1.GetRequest) error {
-	if req.Name == "" {
-		return fmt.Errorf("missing name")
-	}
-	if req.Account == "" {
-		return fmt.Errorf("missing namespace")
-	}
-	return nil
-}
-
-func ValidateListRequest(req *v1.ListRequest) ([]client.ListOption, error) {
-	var listOptions []client.ListOption
-	if req.Account == "" {
-		return listOptions, fmt.Errorf("missing namespace")
-	}
-	listOptions = append(listOptions, client.InNamespace(req.Account))
-	if req.Limit < 0 {
-		return listOptions, fmt.Errorf("invalid limit: should not be negative number")
-	}
-	listOptions = append(listOptions, client.Limit(req.Limit))
-	if req.LabelSelector != "" {
-		selector, err := labels.Parse(req.LabelSelector)
-		if err != nil {
-			return listOptions, fmt.Errorf("invalid LabelSelector: %w", err)
-		}
-		listOptions = append(listOptions, client.MatchingLabelsSelector{
-			Selector: selector,
-		})
-	}
-	if req.Continue != "" {
-		listOptions = append(listOptions, client.Continue(req.Continue))
-	}
-	return listOptions, nil
 }
