+++ pkg/apiserver/internal/v1/region.go
@@ -45,7 +45,7 @@ func NewRegionServiceServer(c client.Client) v1.RegionServiceServer {
 
 func (o regionServer) List(ctx context.Context, req *v1.ListRequest) (res *v1.RegionList, err error) {
 	var listOptions []client.ListOption
-	listOptions, err = util.ValidateListRequest(req)
+	listOptions, err = validateListRequest(req)
 	if err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
@@ -62,7 +62,7 @@ func (o regionServer) List(ctx context.Context, req *v1.ListRequest) (res *v1.Re
 }
 
 func (o regionServer) Get(ctx context.Context, req *v1.GetRequest) (res *v1.Region, err error) {
-	if err = util.ValidateGetRequest(req); err != nil {
+	if err = validateGetRequest(req); err != nil {
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 	region := &catalogv1alpha1.Region{}
