+++ pkg/apiserver/internal/v1/validation.go
@@ -1,0 +1,61 @@
+/*
+Copyright 2020 The KubeCarrier Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1
+
+import (
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/labels"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	v1 "github.com/kubermatic/kubecarrier/pkg/apiserver/api/v1"
+)
+
+func validateGetRequest(req *v1.GetRequest) error {
+	if req.Name == "" {
+		return fmt.Errorf("missing name")
+	}
+	if req.Account == "" {
+		return fmt.Errorf("missing namespace")
+	}
+	return nil
+}
+
+func validateListRequest(req *v1.ListRequest) ([]client.ListOption, error) {
+	var listOptions []client.ListOption
+	if req.Account == "" {
+		return listOptions, fmt.Errorf("missing namespace")
+	}
+	listOptions = append(listOptions, client.InNamespace(req.Account))
+	if req.Limit < 0 {
+		return listOptions, fmt.Errorf("invalid limit: should not be negative number")
+	}
+	listOptions = append(listOptions, client.Limit(req.Limit))
+	if req.LabelSelector != "" {
+		selector, err := labels.Parse(req.LabelSelector)
+		if err != nil {
+			return listOptions, fmt.Errorf("invalid LabelSelector: %w", err)
+		}
+		listOptions = append(listOptions, client.MatchingLabelsSelector{
+			Selector: selector,
+		})
+	}
+	if req.Continue != "" {
+		listOptions = append(listOptions, client.Continue(req.Continue))
+	}
+	return listOptions, nil
+}
